pyproject.toml
[tool.poetry]
name = "tgfp-bin"
version = "1.2.0"
description = "Scripts to help with the running of The Great Football Pool"
authors = ["John Sturgeon <john.sturgeon@me.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10.5,<3.11"
doppler-env = "^0.3.0"
tgfp-lib = "^1.15.0"
hikari = "^2.0.0.dev110"
sentry-sdk = "^1.9.8"
hikari-lightbulb = "^2.2.4"
arrow = "^1.2.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
